Fix some type inconsistencies

+++ kmp_cpu.cpp
@@ -2,9 +2,9 @@
 #include "char_compress.hpp"
 
 void get_fail(const char *pattern, int16_t pattern_length, int16_t *fail) {
-    int candidate = 0;
+    int16_t candidate = 0;
     fail[0] = -1;
-    for (int pos = 1; pos < pattern_length; pos++, candidate++) {
+    for (int16_t pos = 1; pos < pattern_length; pos++, candidate++) {
         char com_pos = get(pattern, pos);
         char com_cand = get(pattern, candidate);
         if (com_pos == com_cand) {
